+++ pynumaflow/errors.py
@@ -1,0 +1,2 @@
+class NoPublicConstructorError(TypeError):
+    """Raise when using ClassName() to create objects while public constructor is not supported"""
